Доработки

+++ Jobs/Connectors/SourceRecord.cs
@@ -3,6 +3,5 @@
 public sealed record SourceRecord<T>(T Value, SourcePosition Position);
 
 public sealed record SinkRecord<T>(T Value);
-    
 
-public sealed record SourcePosition(int Offset );
+public sealed record SourcePosition(long Offset);
